backend/requirements.txt
# ==================== FLEXYOURBRAIN - REQUIREMENTS.TXT ====================
# For Python 3.10+ compatibility

# ==================== BUILD TOOLS ====================
setuptools>=69.0.0
wheel>=0.42.0
pip>=24.0

# ==================== CORE FASTAPI ====================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.6
python-dotenv==1.0.1
aiofiles==24.1.0

# ==================== DATABASE ====================
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
databases==0.9.0

# ==================== AUTHENTICATION & SECURITY ====================
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.7.0
email-validator==2.0.0
bcrypt==4.0.1

# ==================== PYDANTIC ====================
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic[email]==2.10.6

# ==================== AI & QUESTION GENERATION ====================
groq==0.11.0
requests==2.32.3

# ==================== UTILITIES ====================
pytz==2024.2
tqdm==4.67.1
python-magic==0.4.27

# ==================== ASYNC SUPPORT ====================
httpx==0.27.2
anyio==4.6.0